sessions/03_text_embedding/requirements.txt
google-cloud-bigquery[bqstorage,pandas]
google-auth
pandas-gbq
transformers
torch
tqdm
pandas